task22/neuro_control/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
faster-whisper
websockets
PyJWT
requests
aiologger
aiofiles
python-multipart
websocket-client
faster-whisper
langchain
langchain_community
langchain_openai
faiss-cpu
openai
pydub
# audioop-lts




